chore: fixing stuff
