+++ packages/react-storage/src/StorageManager/FileList/FileStatusMessage.tsx
@@ -2,11 +2,8 @@ import React from 'react';
 import classNames from 'classnames';
 
 import { Text, ComponentClassNames } from '@aws-amplify/ui-react';
-import {
-  IconCheck,
-  IconError,
-  classNameModifier,
-} from '@aws-amplify/ui-react/internal';
+import { IconCheck, IconError } from '@aws-amplify/ui-react/internal';
+import { classNameModifier } from '@aws-amplify/ui';
 import { FileStatus } from '../types';
 import { FileStatusMessageProps } from './types';
