+++ packages/react/src/internal.ts
@@ -2,7 +2,6 @@ export * from './hooks/useAuth';
 export * from './hooks/useDataStore';
 export * from './hooks/useStorageURL';
 export * from './primitives/Icon/internal';
-export * from './primitives/shared';
 export {
   UseAuthSignOutAction,
   useAuthSignOutAction,
