+++ packages/react/src/primitives/shared/__tests__/utils.test.tsx
@@ -2,12 +2,8 @@ import { createTheme } from '@aws-amplify/ui';
 import {
   getConsecutiveIntArray,
   strHasLength,
-  classNameModifier,
-  classNameModifierByFlag,
   getCSSVariableIfValueIsThemeKey,
-  Modifiers,
 } from '../utils';
-import { ComponentClassNames } from '../constants';
 
 const theme = createTheme();
 const { tokens } = theme;
@@ -43,39 +39,6 @@ describe('strHasLength:', () => {
 
   it('should return true for strings with a length', () => {
     expect(strHasLength('some string')).toBe(true);
-  });
-});
-
-describe('classNameModifier', () => {
-  const modifiedClassName = 'amplify-alert--modified';
-  const myClass = ComponentClassNames['Alert'];
-  const modifier = 'modified';
-
-  it('should return the modified className with a modifier passed in', () => {
-    expect(classNameModifier(myClass, modifier)).toEqual(modifiedClassName);
-  });
-
-  it('should return empty string without a modifier passed in', () => {
-    // force undefined to be Modifiers type for exhaustive edge case test
-    expect(
-      classNameModifier(myClass, undefined as unknown as Modifiers)
-    ).toEqual('');
-  });
-});
-
-describe('classNameModifierByFlag', () => {
-  const modifiedClassName = 'amplify-alert--modified';
-  const myClass = ComponentClassNames['Alert'];
-  const modifier = 'modified';
-
-  it('should return the modified className with a true flag value passed in', () => {
-    expect(classNameModifierByFlag(myClass, modifier, true)).toEqual(
-      modifiedClassName
-    );
-  });
-
-  it('should return empty string with a false flag value passed in', () => {
-    expect(classNameModifierByFlag(myClass, modifier, false)).toEqual('');
   });
 });
