+++ packages/react/src/primitives/shared/index.ts
@@ -1,2 +1,1 @@
 export { ComponentClassNames, ComponentClassObject } from './constants';
-export { classNameModifier, classNameModifierByFlag } from './utils';
