+++ packages/react/src/primitives/shared/utils.ts
@@ -1,7 +1,8 @@
 import {
   cssNameTransform,
   isDesignToken,
-  ComponentClassName,
+  classNameModifier,
+  classNameModifierByFlag,
   WebTheme,
 } from '@aws-amplify/ui';
 
@@ -29,38 +30,6 @@ export const getConsecutiveIntArray = (
 ): number[] => {
   const length = end - start + 1;
   return Array.from({ length }, (_, idx) => idx + start);
-};
-
-export type Modifiers = string | number | null;
-
-/**
- * This helper function creates modifier class names that are used for our flat BEM styling
- * it takes in a base and modifier and returns the modified class if a modifier was passed in and null otherwise
- * @param base The base class of the output
- * @param modifier The modifier to add onto the base
- * @returns the modified class name or empty string
- */
-export const classNameModifier = (
-  base: ComponentClassName,
-  modifier?: Modifiers
-): string => {
-  return modifier ? `${base}--${modifier}` : '';
-};
-
-/**
- * This helper function creates modified class names that are used for our flat BEM styling
- * it takes in a base, modifier, and flag and returns the modified class name if the flag is true and null if the flag is false
- * @param base
- * @param modifier
- * @param flag
- * @returns the modified class name or empty string
- */
-export const classNameModifierByFlag = (
-  base: ComponentClassName,
-  modifier: Modifiers,
-  flag?: boolean
-): string => {
-  return flag ? `${base}--${modifier}` : '';
 };
 
 /**
@@ -111,3 +80,5 @@ export const getCSSVariableIfValueIsThemeKey = <Value = unknown>(
 
   return value;
 };
+
+export { classNameModifier, classNameModifierByFlag };
